refactor: update package structure
Add data package to group together all repository and data source classes.
