+++ app/src/main/java/com/washuTechnologies/merced/data/database/SpaceXDatabase.kt
@@ -1,11 +1,11 @@
-package com.washuTechnologies.merced.database
+package com.washuTechnologies.merced.data.database
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
-import com.washuTechnologies.merced.api.launches.RocketLaunch
-import com.washuTechnologies.merced.database.converters.StringArrayConverter
-import com.washuTechnologies.merced.database.launches.LaunchesLocalDatasource
+import com.washuTechnologies.merced.data.launches.model.RocketLaunch
+import com.washuTechnologies.merced.data.database.converters.StringArrayConverter
+import com.washuTechnologies.merced.data.launches.datasources.LaunchesLocalDatasource
 
 @Database(entities = [RocketLaunch::class], version = 1)
 @TypeConverters(StringArrayConverter::class)
