+++ app/src/main/java/com/washuTechnologies/merced/data/launches/datasources/LaunchesLocalDatasource.kt
@@ -1,10 +1,10 @@
-package com.washuTechnologies.merced.database.launches
+package com.washuTechnologies.merced.data.launches.datasources
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.washuTechnologies.merced.api.launches.RocketLaunch
+import com.washuTechnologies.merced.data.launches.model.RocketLaunch
 
 @Dao
 interface LaunchesLocalDatasource {
