+++ app/src/test/java/com/washuTechnologies/merced/intergrationtest/ui/launchlist/LaunchListViewModelIntegrationTest.kt
@@ -1,7 +1,7 @@
 package com.washuTechnologies.merced.intergrationtest.ui.launchlist
 
 import android.accounts.NetworkErrorException
-import com.washuTechnologies.merced.api.launches.LaunchesRemoteDatasource
+import com.washuTechnologies.merced.data.launches.datasources.LaunchesRemoteDatasource
 import com.washuTechnologies.merced.ui.launchlist.LaunchListUiState
 import com.washuTechnologies.merced.ui.launchlist.LaunchListViewModel
 import com.washuTechnologies.merced.util.MockDatasourceHelper
